pyproject.toml
[project]
name = "SlideSage-backend"
version = "0.1.0"
description = "Backend for SlideSage"
requires-python = ">=3.10"
dependencies = [
	"flask==3.0.2",
	"flask-cors==4.0.0",
	"python-dotenv==1.0.0",
	"litellm>=0.0.0",
	"python-pptx==0.6.23",
	"pydantic==2.5.0",
	"watchdog==3.0.0",
	"google-auth>=2.0.0",
	"google-cloud-aiplatform>=1.30.0",
	"vercel>=0.2.1",
]
